# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Absensi API
  version: 0.1.0-draft
  description: >
    Draft API contract for Phase 1. Provisional pending live DB dump review
    (docs/DECISIONS.md D-12). Unifies the legacy PresensiController and
    AbsensiController/EOS flows into one check-in/out flow per D-8.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /me:
    get:
      tags: [profile]
      summary: Current user's own profile (employee or admin)
      security:
        - bearerAuth: [employee, admin]
      responses:
        "200": { description: OK }
    patch:
      tags: [profile]
      summary: Update own editable fields (currently just phone)
      security:
        - bearerAuth: [employee, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
      responses:
        "200": { description: OK }
...
  /attendance/check-in:
    post:
      tags: [attendance]
      summary: >
        Unified check-in (replaces legacy /presensi/store, /presensi/wfhstore,
        /absensi/store -- D-8). Server determines in vs out based on today's
        (or a still-open overnight, D-14) attendance row. Server-side geofencing
        enforced per D-1 (100m default, from office_locations, not hardcoded);
        rejects the request if outside radius, unless is_wfh is true or the
        employee has an approved field_assignment (dinas luar, D-21) for today.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude, photo, is_wfh]
              properties:
                latitude: { type: number, format: double }
                longitude: { type: number, format: double }
                photo:
                  type: string
                  format: byte
                  description: Base64-encoded selfie photo (validated server-side, unlike legacy)
                is_wfh: { type: boolean }
      responses:
        "200":
          description: Check-in/out recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendance" }
        "409":
          description: Duplicate/cooldown violation (idempotency, D-9)
        "422":
          description: Outside geofence radius (D-1) or validation failure
...
  /admin/leave-requests/{id}/review:
    post:
      tags: [leave, admin]
      summary: Approve or reject a leave request (admin, D-15)
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveRequest" }

  # ------------------------------------------------------------------ tasks
...
  /admin/reports/recap/export:
    get:
      tags: [reports, admin]
      summary: >
        Same recap grid as GET /admin/reports/recap, streamed as an .xlsx
        file -- calls the identical recap.Service.Generate so the export can
        never drift from what's shown on screen (single source of truth).
      security:
        - bearerAuth: [admin]
      parameters:
        - name: employee_id
          in: query
          schema: { type: integer }
        - name: month
          in: query
          required: true
          schema: { type: integer }
        - name: year
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: {}
...
  /admin/employees/{id}:
    put:
      tags: [employees, admin]
      summary: >
        Update an employee. Password is only rehashed when a new one is
        supplied in the request body; omitting it leaves the existing hash
        untouched (does not replicate the legacy reset-on-edit bug, D-3).
        Also used to deactivate/reactivate via is_active.
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeInput" }
      responses:
        "200": { description: OK }
        "404": { description: Employee not found }
...
  /admin/config/field-assignments:
    get:
      tags: [config, admin]
      summary: List "dinas luar" (off-site work) pre-approvals (D-21)
      security:
        - bearerAuth: [admin]
      parameters:
        - name: employee_id
          in: query
          schema: { type: integer }
      responses:
        "200": { description: OK }
    post:
      tags: [config, admin]
      summary: >
        Pre-approve an employee for off-site work on a given date. Check-in on
        that date bypasses the office_locations radius check (D-1/D-21).
      security:
        - bearerAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employee_id, work_date]
              properties:
                employee_id: { type: integer }
                work_date: { type: string, format: date }
                note: { type: string }
      responses:
        "201": { description: Created }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    LeaveType:
      type: string
      enum: [izin, sakit]
    LeaveStatus:
      type: string
      enum: [pending, approved, rejected]
    LeaveRequest:
      type: object
      properties:
        id: { type: integer }
        employee_id: { type: integer }
        type: { $ref: "#/components/schemas/LeaveType" }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        reason: { type: string }
        status: { $ref: "#/components/schemas/LeaveStatus" }
        reviewed_by: { type: integer, nullable: true }
        reviewed_at: { type: string, format: date-time, nullable: true }
    Attendance:
      type: object
      properties:
        id: { type: integer }
        employee_id: { type: integer }
        work_date: { type: string, format: date }
        cycle_number: { type: integer, description: "Supports multiple check-in/out cycles per day, D-23" }
        shift_id: { type: integer, nullable: true }
        is_wfh: { type: boolean }
        check_in_at: { type: string, format: date-time, nullable: true }
        check_in_distance_m: { type: integer, nullable: true }
        is_late: { type: boolean, nullable: true }
        check_out_at: { type: string, format: date-time, nullable: true }
        check_out_distance_m: { type: integer, nullable: true }
        is_early_leave: { type: boolean, nullable: true }
        status: { type: string, enum: [open, closed, flagged_no_checkout] }
...
    EmployeeInput:
      type: object
      required: [nik, full_name]
      properties:
        nik: { type: string }
        full_name: { type: string }
        password:
          type: string
          description: Required on create. Optional on update -- omit to keep the existing password.
        department_id: { type: integer, nullable: true }
        position: { type: string }
        phone: { type: string }
        is_active: { type: boolean, default: true }
...